# backend/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Middle Monitor API",
    "version": "1.0.0",
    "description": "REST API of the Middle Monitor observability platform.\n\nAuthentication is a bearer token: a personal API token (Settings, API tokens)\nor an organization API key. Agent install tokens are a separate credential and\nare only accepted on the /agents endpoints, through the X-Install-Token header.\n\nList endpoints accept limit and offset and return the total in X-Total-Count.",
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.middlemonitor.io",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Personal token or organization API key"
      },
      "installToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Install-Token",
        "description": "Agent install token, accepted only on the agent endpoints"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request is invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/v1/auth/me": {
      "delete": {
        "operationId": "handleDeleteAccount",
        "summary": "Delete account",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "description": "Delete your account. Body {\"password\": \"...\"}. Organizations you are the only member of are deleted with it; 409 if you are the last admin of one that has other members.",
        "security": []
      },
      "get": {
        "operationId": "handleGetMe",
        "summary": "Get me",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": []
      }
    },
...
    "/api/v1/contact": {
      "post": {
        "operationId": "contactLimiter.Middleware",
        "summary": "Contact limiter.middleware",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/v1/organizations/{org_slug}": {
      "delete": {
        "operationId": "handleDeleteOrganization",
        "summary": "Delete organization",
        "tags": [
          "root"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "description": "Delete the organization and all its data (admin only). Body {\"confirm\": \"<org_slug>\"}. Cancels its subscription; members of other organizations keep their account.",
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Organization slug"
            }
          }
        ]
      },
      "get": {
        "operationId": "handleGetOrganization",
        "summary": "Get organization",
        "tags": [
          "root"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Organization slug"
            }
          }
        ]
      },
      "put": {
        "operationId": "handleUpdateOrganization",
        "summary": "Update organization",
        "tags": [
          "root"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Organization slug"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Opt-in: a retry carrying the same key replays the first answer (Idempotent-Replay: true) instead of creating a second row. 409 with code idempotency_in_progress while the first call is still running."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/status": {
      "get": {
        "operationId": "statusLimiter.Middleware",
        "summary": "Status limiter.middleware",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": []
      }
    },
    "/api/v1/unsubscribe": {
      "post": {
        "operationId": "unsubLimiter.Middleware",
        "summary": "Unsub limiter.middleware",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": []
      }
    },
...
  }
}